+++ hans_cute_driver/src/hans_cute_driver.cpp
@@ -1,9 +1,9 @@
 #include "hans_cute_driver/hans_cute_driver.h"
 
-const SamplePacket HansCuteRobot::ServoDriver::DXL_PACKET{ std::vector<uint8_t>({ 0xFF, 0xFF }), 2, 3, 4 };
+const SamplePacket HansCuteRobot::ServoDriver::DXL_PACKET{std::vector<uint8_t>({0xFF, 0xFF}), 2, 3, 4};
 
 HansCuteRobot::ServoDriver::ServoDriver(const std::string port, const long baudrate)
-  : HansCuteRobot::ServoSerialComms(port, baudrate)
+    : HansCuteRobot::ServoSerialComms(port, baudrate)
 {
   // Update paket format at initialisation
   setSamplePacket(HansCuteRobot::ServoDriver::DXL_PACKET);
@@ -19,21 +19,21 @@ HansCuteRobot::ServoDriver::~ServoDriver()
 //====================================================================//
 
 // These can be left until the end as they are relatively unimportant
-bool HansCuteRobot::ServoDriver::setID(const uint8_t& old_id, const uint8_t& new_id)
+bool HansCuteRobot::ServoDriver::setID(const uint8_t &old_id, const uint8_t &new_id)
 {
   return true;
 }
-bool HansCuteRobot::ServoDriver::setBaudrate(const uint8_t& servo_id, const long& baudrate)
+bool HansCuteRobot::ServoDriver::setBaudrate(const uint8_t &servo_id, const long &baudrate)
 {
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::setReturnDelayTime(const uint8_t& servo_id, const unsigned int& delay_time)
+bool HansCuteRobot::ServoDriver::setReturnDelayTime(const uint8_t &servo_id, const unsigned int &delay_time)
 {
   return true;
 }
-bool HansCuteRobot::ServoDriver::setAngleLimits(const uint8_t& servo_id, const unsigned int& min_limit,
-                                                const unsigned int& max_limit)
+bool HansCuteRobot::ServoDriver::setAngleLimits(const uint8_t &servo_id, const unsigned int &min_limit,
+                                                const unsigned int &max_limit)
 {
   std::vector<uint8_t> raw_limit = HansCuteRobot::AngleLimits::getRawData(min_limit, max_limit);
   std::vector<uint8_t> returned_data;
@@ -42,12 +42,12 @@ bool HansCuteRobot::ServoDriver::setAngleLimits(const uint8_t& servo_id, const u
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::setVoltageLimits(const uint8_t &servo_id, const VoltageLimits& voltage_limits)
+bool HansCuteRobot::ServoDriver::setVoltageLimits(const uint8_t &servo_id, const VoltageLimits &voltage_limits)
 {
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::setMaxTorque(const uint8_t& servo_id, const unsigned int& max_torque)
+bool HansCuteRobot::ServoDriver::setMaxTorque(const uint8_t &servo_id, const unsigned int &max_torque)
 {
   std::vector<uint8_t> raw_limit = HansCuteRobot::TorqueLimit::getRawData(max_torque);
   std::vector<uint8_t> returned_data;
@@ -55,7 +55,7 @@ bool HansCuteRobot::ServoDriver::setMaxTorque(const uint8_t& servo_id, const uns
   write(servo_id, (uint8_t)ControlTableConstant::MAX_TORQUE_L, raw_limit, returned_data, timestamp);
   for (uint8_t data : returned_data)
   {
-    std::cout << (int)data << std::endl; 
+    std::cout << (int)data << std::endl;
   }
   return true;
 }
@@ -63,9 +63,9 @@ bool HansCuteRobot::ServoDriver::setMaxTorque(const uint8_t& servo_id, const uns
 //===============================================================//
 // These functions can send multiple commands to a single servo  //
 //===============================================================//
-bool HansCuteRobot::ServoDriver::setTorqueEnable(const uint8_t& servo_id, const bool& enabled)
+bool HansCuteRobot::ServoDriver::setTorqueEnable(const uint8_t &servo_id, const bool &enabled)
 {
-  std::vector<uint8_t> data({ (uint8_t)enabled });
+  std::vector<uint8_t> data({(uint8_t)enabled});
   std::vector<uint8_t> returned_data;
   unsigned long timestamp;
   write(servo_id, (uint8_t)ControlTableConstant::TORQUE_ENABLE, data, returned_data, timestamp);
@@ -92,7 +92,7 @@ bool HansCuteRobot::ServoDriver::setPGain()
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::setAcceleration(const uint8_t& servo_id, const unsigned int& acceleration)
+bool HansCuteRobot::ServoDriver::setAcceleration(const uint8_t &servo_id, const unsigned int &acceleration)
 {
   std::vector<uint8_t> raw_data = HansCuteRobot::ServoAcceleration::getRawData(acceleration);
   std::vector<uint8_t> returned_data;
@@ -100,7 +100,7 @@ bool HansCuteRobot::ServoDriver::setAcceleration(const uint8_t& servo_id, const 
   write(servo_id, (uint8_t)ControlTableConstant::GOAL_ACCELERATION, raw_data, returned_data, timestamp);
   return true;
 }
-bool HansCuteRobot::ServoDriver::setPosition(const uint8_t& servo_id, const unsigned int& position)
+bool HansCuteRobot::ServoDriver::setPosition(const uint8_t &servo_id, const unsigned int &position)
 {
   std::vector<uint8_t> raw_postion = HansCuteRobot::ServoPosition::getRawData(position);
   std::vector<uint8_t> returned_data;
@@ -108,7 +108,7 @@ bool HansCuteRobot::ServoDriver::setPosition(const uint8_t& servo_id, const unsi
   write(servo_id, (uint8_t)ControlTableConstant::GOAL_POSITION_L, raw_postion, returned_data, timestamp);
   return true;
 }
-bool HansCuteRobot::ServoDriver::setSpeed(const uint8_t& servo_id, const unsigned int& speed)
+bool HansCuteRobot::ServoDriver::setSpeed(const uint8_t &servo_id, const unsigned int &speed)
 {
   std::vector<uint8_t> raw_speed = HansCuteRobot::ServoSpeed::getRawData(speed);
   std::vector<uint8_t> returned_data;
@@ -117,7 +117,7 @@ bool HansCuteRobot::ServoDriver::setSpeed(const uint8_t& servo_id, const unsigne
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::setTorqueLimit(const uint8_t& servo_id, const unsigned int& torque_limit)
+bool HansCuteRobot::ServoDriver::setTorqueLimit(const uint8_t &servo_id, const unsigned int &torque_limit)
 {
   std::vector<uint8_t> raw_torque = HansCuteRobot::TorqueLimit::getRawData(torque_limit);
   std::vector<uint8_t> returned_data;
@@ -155,8 +155,8 @@ bool HansCuteRobot::ServoDriver::setMultiComplianceSlope()
 }
 
 // Position value ranges from 0 -> 4095 (0xFFF), unit is 0.088 degree
-bool HansCuteRobot::ServoDriver::setMultiPosition(const std::vector<unsigned int>& servo_ids,
-                                                  const std::vector<unsigned int>& positions)
+bool HansCuteRobot::ServoDriver::setMultiPosition(const std::vector<unsigned int> &servo_ids,
+                                                  const std::vector<unsigned int> &positions)
 {
   // Prepare data for sync_write
   std::vector<std::vector<uint8_t>> data_lists;
@@ -171,8 +171,8 @@ bool HansCuteRobot::ServoDriver::setMultiPosition(const std::vector<unsigned int
   syncWrite((uint8_t)ControlTableConstant::GOAL_POSITION_L, data_lists);
   return true;
 }
-bool HansCuteRobot::ServoDriver::setMultiSpeed(const std::vector<unsigned int>& servo_ids,
-                                               const std::vector<unsigned int>& speeds)
+bool HansCuteRobot::ServoDriver::setMultiSpeed(const std::vector<unsigned int> &servo_ids,
+                                               const std::vector<unsigned int> &speeds)
 {
   // Prepare data for sync_write
   std::vector<std::vector<uint8_t>> data_lists;
@@ -188,14 +188,14 @@ bool HansCuteRobot::ServoDriver::setMultiSpeed(const std::vector<unsigned int>& 
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::setMultiTorqueLimit(const std::vector<unsigned int>& servo_ids,
-                                                     const std::vector<unsigned int>& torque_limits)
+bool HansCuteRobot::ServoDriver::setMultiTorqueLimit(const std::vector<unsigned int> &servo_ids,
+                                                     const std::vector<double> &torque_limits)
 {
   return true;
 }
-bool HansCuteRobot::ServoDriver::setMultiPositionAndSpeed(const std::vector<unsigned int>& servo_ids,
-                                                          const std::vector<unsigned int>& positions,
-                                                          const std::vector<unsigned int>& speeds)
+bool HansCuteRobot::ServoDriver::setMultiPositionAndSpeed(const std::vector<unsigned int> &servo_ids,
+                                                          const std::vector<unsigned int> &positions,
+                                                          const std::vector<unsigned int> &speeds)
 {
   std::vector<std::vector<uint8_t>> data_lists;
   for (int indx = 0; indx < servo_ids.size(); indx++)
@@ -216,7 +216,7 @@ bool HansCuteRobot::ServoDriver::setMultiPositionAndSpeed(const std::vector<unsi
 // Servo status access functions //
 //===============================//
 
-bool HansCuteRobot::ServoDriver::getModelNumber(const int& servo_id, unsigned int& model_number)
+bool HansCuteRobot::ServoDriver::getModelNumber(const int &servo_id, unsigned int &model_number)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -236,7 +236,7 @@ bool HansCuteRobot::ServoDriver::getReturnDelayTime()
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::getAngleLimits(const int& servo_id, HansCuteRobot::AngleLimits& angle_limit)
+bool HansCuteRobot::ServoDriver::getAngleLimits(const int &servo_id, HansCuteRobot::AngleLimits &angle_limit)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -248,7 +248,7 @@ bool HansCuteRobot::ServoDriver::getAngleLimits(const int& servo_id, HansCuteRob
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::getTorqueEnabled(const int& servo_id, bool& enabled)
+bool HansCuteRobot::ServoDriver::getTorqueEnabled(const int &servo_id, bool &enabled)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -259,7 +259,7 @@ bool HansCuteRobot::ServoDriver::getTorqueEnabled(const int& servo_id, bool& ena
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::getMaxTorque(const int& servo_id, unsigned int& torque)
+bool HansCuteRobot::ServoDriver::getMaxTorque(const int &servo_id, unsigned int &torque)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -275,7 +275,7 @@ bool HansCuteRobot::ServoDriver::getVoltageLimits()
 {
   return true;
 }
-bool HansCuteRobot::ServoDriver::getPosition(const int& servo_id, unsigned int& position)
+bool HansCuteRobot::ServoDriver::getPosition(const int &servo_id, unsigned int &position)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -286,7 +286,7 @@ bool HansCuteRobot::ServoDriver::getPosition(const int& servo_id, unsigned int& 
   position = (unsigned int)HansCuteRobot::ServoPosition::getData(response);
   return true;
 }
-bool HansCuteRobot::ServoDriver::getSpeed(const int& servo_id, unsigned int& speed)
+bool HansCuteRobot::ServoDriver::getSpeed(const int &servo_id, unsigned int &speed)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -307,7 +307,7 @@ bool HansCuteRobot::ServoDriver::getCurrent()
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::getLock(const int& servo_id, bool& lock)
+bool HansCuteRobot::ServoDriver::getLock(const int &servo_id, bool &lock)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
@@ -319,7 +319,7 @@ bool HansCuteRobot::ServoDriver::getLock(const int& servo_id, bool& lock)
   return true;
 }
 
-bool HansCuteRobot::ServoDriver::getFeedback(const int& servo_id, ServoFeedback& feedback)
+bool HansCuteRobot::ServoDriver::getFeedback(const int &servo_id, ServoFeedback &feedback)
 {
   std::vector<uint8_t> response;
   unsigned long timestamp = 0;
